Move /me to / and delete /me
