+++ apps/web/src/components/verification/form/index.tsx
@@ -45,7 +45,7 @@ export default function VerificationForm() {
         setLoading(false);
         setFormError(result.message);
       } else {
-        router.push('/me');
+        router.push('/');
       }
     } catch (error) {
       setLoading(false);
